+++ crates/ruma-client-api/src/rendezvous/create_rendezvous_session.rs
@@ -2,8 +2,9 @@
 //!
 //! Create a rendezvous session.
 
-pub mod unstable {
-    //! `msc4108` ([MSC])
+#[cfg(feature = "unstable-msc4108")]
+pub mod unstable_msc4108 {
+    //! `unstable/org.matrix.msc4108` ([MSC])
     //!
     //! [MSC]: https://github.com/matrix-org/matrix-spec-proposals/pull/4108
 
@@ -27,7 +28,7 @@ pub mod unstable {
         }
     }
 
-    /// Request type for the `POST` `rendezvous` endpoint.
+    /// Request type for the `POST` `rendezvous` endpoint from the 2024 version of MSC4108.
     #[derive(Debug, Default, Clone)]
     #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
     pub struct Request {
@@ -112,7 +113,7 @@ pub mod unstable {
         }
     }
 
-    /// Response type for the `POST` `rendezvous` endpoint.
+    /// Response type for the `POST` `rendezvous` endpoint from the 2024 version of MSC4108.
     #[derive(Debug, Clone)]
     #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
     pub struct Response {
@@ -200,6 +201,63 @@ pub mod unstable {
                 .header(EXPIRES, expires)
                 .header(LAST_MODIFIED, last_modified)
                 .body(body)?)
+        }
+    }
+}
+
+#[cfg(feature = "unstable-msc4388")]
+pub mod unstable_msc4388 {
+    //! `unstable/io.element.msc4388` ([MSC])
+    //!
+    //! [MSC]: https://github.com/matrix-org/matrix-spec-proposals/pull/4388
+    use std::time::Duration;
+
+    use ruma_common::{
+        api::{auth_scheme::AccessTokenOptional, request, response},
+        metadata,
+    };
+
+    metadata! {
+        method: POST,
+        rate_limited: true,
+        authentication: AccessTokenOptional,
+        history: {
+            unstable("io.element.msc4388") => "/_matrix/client/unstable/io.element.msc4388/rendezvous",
+        }
+    }
+
+    /// Request type for the `POST` `rendezvous` endpoint.
+    #[request(error = crate::Error)]
+    pub struct Request {
+        /// Data up to maximum size allowed by the server.
+        pub data: String,
+    }
+
+    impl Request {
+        /// Creates a new `Request` with the given content.
+        pub fn new(data: String) -> Self {
+            Self { data }
+        }
+    }
+
+    /// Response type for the `POST` `rendezvous` endpoint.
+    #[response(error = crate::Error)]
+    pub struct Response {
+        /// The ID of the created rendezvous session.
+        pub id: String,
+
+        /// The initial sequence token for the session.
+        pub sequence_token: String,
+
+        /// The time remaining in milliseconds until the session expires.
+        #[serde(with = "ruma_common::serde::duration::ms", rename = "expires_in_ms")]
+        pub expires_in: Duration,
+    }
+
+    impl Response {
+        /// Creates a new `Response` with the given content.
+        pub fn new(id: String, sequence_token: String, expires_in: Duration) -> Self {
+            Self { id, sequence_token, expires_in }
         }
     }
 }